requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.5
colorama==0.4.6
colorthief==0.2.1
comm==0.1.3
contourpy==1.1.0
convcolors==2.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==4.4.2
executing==1.2.0
extcolors==1.0.0
fast-colorthief==0.0.4
filedate==2.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.41.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1.1
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.9.1
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
Pygments==2.15.1
pyparsing==3.0.9
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.136
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
